pyproject.toml
[project]
name = "youtube-to-text"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "assemblyai>=0.37.0",
    "audioop-lts>=0.2.1",
    "pydantic-settings>=2.8.1",
    "pydub>=0.25.1",
    "python-dotenv>=1.0.1",
    "pytubefix>=8.12.2",
]
